% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSDAG.R
\name{emvs}
\alias{emvs}
\title{VSDAG}
\usage{
emvs(seed, n, p, x)
}
\arguments{
\item{seed}{set a seed}

\item{n}{sample size of the observed data $x$}

\item{p}{dimension of the observed data $x$}

\item{x}{observed data $x$}
}
\value{
adjacent matrix gamma and weight matrix beta
}
\description{
VSDAG
}
\examples{
emvs(1,100,5,matrix(rnorm(500),nrow=5,ncol=100))
}
